Upgrades to Rust 1.65.0 (#28741)

+++ program-runtime/src/compute_budget.rs
@@ -594,7 +594,7 @@ mod tests {
             )),
             ComputeBudget {
                 compute_unit_limit: MAX_COMPUTE_UNIT_LIMIT as u64,
-                heap_size: Some(MIN_HEAP_FRAME_BYTES as usize),
+                heap_size: Some(MIN_HEAP_FRAME_BYTES),
                 ..ComputeBudget::default()
             },
             false,
@@ -616,7 +616,7 @@ mod tests {
             &[
                 Instruction::new_with_bincode(Pubkey::new_unique(), &0_u8, vec![]),
                 ComputeBudgetInstruction::request_heap_frame(MIN_HEAP_FRAME_BYTES as u32),
-                ComputeBudgetInstruction::request_heap_frame(MAX_HEAP_FRAME_BYTES as u32),
+                ComputeBudgetInstruction::request_heap_frame(MAX_HEAP_FRAME_BYTES),
             ],
             Err(TransactionError::DuplicateInstruction(2)),
             ComputeBudget::default()
