+++ programs/bpf_loader/src/syscalls/mod.rs
@@ -649,7 +649,7 @@ declare_syscall!(
         };
         *result = if free_addr == 0 {
             match allocator.alloc(layout) {
-                Ok(addr) => Ok(addr as u64),
+                Ok(addr) => Ok(addr),
                 Err(_) => Ok(0),
             }
         } else {
@@ -1931,7 +1931,7 @@ declare_syscall!(
                     translate_slice_mut::<u8>(
                         memory_mapping,
                         data_addr,
-                        result_header.data_len as u64,
+                        result_header.data_len,
                         invoke_context.get_check_aligned(),
                         invoke_context.get_check_size(),
                     ),
@@ -1941,7 +1941,7 @@ declare_syscall!(
                     translate_slice_mut::<AccountMeta>(
                         memory_mapping,
                         accounts_addr,
-                        result_header.accounts_len as u64,
+                        result_header.accounts_len,
                         invoke_context.get_check_aligned(),
                         invoke_context.get_check_size(),
                     ),
