+++ runtime/src/storable_accounts.rs
@@ -163,7 +163,7 @@ pub mod tests {
                         raw.push((
                             pk,
                             AccountSharedData::create(
-                                ((entry as u64) * starting_slot) as u64,
+                                (entry as u64) * starting_slot,
                                 Vec::default(),
                                 Pubkey::default(),
                                 false,
